file reformat & gitignore
